Moved base CQRS entity manager to `Storage` NS
