+++ src/Module/WpBookingsCqrsModule.php
@@ -9,12 +9,11 @@ use Dhii\Exception\InternalException;
 use Dhii\Output\PlaceholderTemplateFactory;
 use Dhii\Util\Normalization\NormalizeArrayCapableTrait;
 use Dhii\Util\String\StringableInterface as Stringable;
-use Exception;
 use mysqli;
 use Psr\Container\ContainerInterface;
 use Psr\EventManager\EventInterface;
 use Psr\EventManager\EventManagerInterface;
-use RebelCode\EddBookings\Logic\Module\BaseCqrsEntityManager;
+use RebelCode\EddBookings\Logic\Storage\BaseCqrsEntityManager;
 use RebelCode\Modular\Module\AbstractBaseModule;
 use RebelCode\Storage\Resource\WordPress\Wpdb\BookingsSelectResourceModel;
 use RebelCode\Storage\Resource\WordPress\Wpdb\BookingStatusWpdbSelectResourceModel;
